finalapp/requirements.txt
flask
flask-wtf
mysql-connector-python
flask-login
flask-security
sqlalchemy
